python/requirements.txt
Flask
requests
gunicorn
google-auth
